+++ test/table/test_table_extraction.py
@@ -93,8 +93,8 @@ if sys.platform not in ("cygwin", "win32"):
 def test_tabula_extract_simple_table(filename):
     dataframes = tabula.read_pdf(HERE / filename, pages="all")
     assert len(dataframes) == 1
-    for df in dataframes:
-        df = df.astype({"Age": str})
+    for df_raw in dataframes:
+        df = df_raw.astype({"Age": str})
         assert_frame_equal(df, TABLE_DATA_AS_DF, check_names=False)
 
 
@@ -108,8 +108,8 @@ def test_tabula_extract_simple_table(filename):
 def test_tabula_extract_two_tables(filename):
     dataframes = tabula.read_pdf(HERE / filename, pages="all")
     assert len(dataframes) == 2
-    for df in dataframes:
-        df = df.astype({"Age": str})
+    for df_raw in dataframes:
+        df = df_raw.astype({"Age": str})
         assert_frame_equal(df, TABLE_DATA_AS_DF, check_names=False)